# api/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Keystone Core REST API
  description: |
    REST API for Keystone Core - a runtime infrastructure control plane.

    This specification covers all HTTP REST endpoints exposed by Keystone Core components.
    The primary control plane operations use gRPC (see proto files), but HTTP REST endpoints
    are available for cluster management, health checks, file operations, and more.

    ## Authentication

    Keystone Core supports multiple authentication methods:
    - **API Key**: Bearer token in `Authorization` header
    - **mTLS**: Client certificate authentication
    - **JWT**: JSON Web Token authentication

    ## Rate Limiting

    API endpoints are subject to rate limiting. When rate limited, the API returns
    HTTP 429 with `Retry-After` header indicating when to retry.

    ## Versioning

    All REST APIs use the `/api/v1/` prefix. Breaking changes will increment the version.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Keystone Core Team
    url: https://github.com/shawnbutts/keystone-core
servers:
  - url: http://localhost:9090
    description: Local development server
  - url: https://{control-plane-host}:9443
    description: Production server (HTTPS)
    variables:
      control-plane-host:
        default: control-plane.example.com
        description: Control plane hostname
...
security:
  - bearerAuth: []
  - mtls: []
paths:
...
  /api/status:
    get:
      tags:
        - Health
      summary: Server status
      description: |
        Returns detailed server status including version, uptime,
        memory usage, goroutine count, and connected agents.
      operationId: serverStatus
      responses:
        '200':
          description: Server status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # Cluster Management (handler not yet wired into kscore-server)
  # ============================================
...
  /api/v1/agents/{id}:
    get:
      tags:
        - Agents
      summary: Get agent details
      description: Returns detailed information about a specific agent.
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          description: Agent ID
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # Execution
  # ============================================
  /api/v1/exec:
    post:
      tags:
        - Execution
      summary: Execute command
      description: |
        Executes a command on one or more target agents. Supports synchronous
        execution (waits for results) and async mode (returns job ID immediately).
      operationId: executeCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/jobs:
    get:
      tags:
        - Execution
      summary: List jobs
      description: Returns a paginated list of command execution jobs.
      operationId: listJobs
      parameters:
        - name: agent_id
          in: query
          description: Filter by agent ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by job status
          schema:
            type: string
            enum: [pending, running, success, failed, timeout, cancelled]
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
        - name: order
          in: query
          description: Sort order (asc or desc)
          schema:
            type: string
            enum: [asc, desc]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication
    mtls:
      type: mutualTLS
      description: Mutual TLS client certificate authentication
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 100
        maximum: 1000
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
...
    ServerStatus:
      type: object
      properties:
        version:
          type: string
        buildDate:
          type: string
          format: date-time
        goVersion:
          type: string
        uptime:
          type: string
        memoryUsage:
          type: integer
          format: int64
          description: Memory usage in bytes
        goroutines:
          type: integer
        connectedAgents:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
        hostname:
          type: string
        os:
          type: string
        arch:
          type: string
        agent_version:
          type: string
        platform_version:
          type: string
        status:
          type: string
          enum: [connected, disconnected, unknown]
        labels:
          type: object
          additionalProperties:
            type: string
        ip_addresses:
          type: array
          items:
            type: string
        ipv4_addresses:
          type: array
          items:
            type: string
        ipv6_addresses:
          type: array
          items:
            type: string
        is_dual_stack:
          type: boolean
        registered_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
...
    ExecuteRequest:
      type: object
      required:
        - command
      properties:
        targets:
          type: array
          items:
            type: string
          description: Target agent IDs
        target:
          type: string
          description: Single target agent ID or expression
        command:
          type: string
          description: Command to execute
        args:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
        working_dir:
          type: string
        user:
          type: string
        shell:
          type: string
        timeout:
          type: integer
          description: Timeout in seconds (default 60)
        async:
          type: boolean
          description: Run asynchronously (return job ID immediately)
    ExecuteResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, dispatched, completed, failed]
        targets:
          type: array
          items:
            type: string
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ExecResult'
        created_at:
          type: string
          format: date-time
    ExecResult:
      type: object
      properties:
        agent_id:
          type: string
        status:
          type: string
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        error:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          format: int64
    JobResponse:
      type: object
      properties:
        id:
          type: string
        agent_id:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, running, success, failed, timeout, cancelled]
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        error:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          format: int64
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        retrieved_at:
          type: string
          format: date-time
...